ml/requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.4
prophet==1.1.5
scikit-learn==1.3.2
fastapi==0.104.1
uvicorn==0.24.0
apache-airflow[postgres]==2.8.1
requests==2.31.0